% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPVE.cov.R
\name{calcPVE.cov}
\alias{calcPVE.cov}
\title{Calculating PVE from mixed-model's results}
\usage{
calcPVE.cov(x)
}
\arguments{
\item{x}{A output of \code{neiEM3.cov()}.}
}
\value{
Proportion of phenotypic variation explained (PVE) by direct genetic effects (DGEs), indirect genetic effects (IGEs), and their covariance (cov).
}
\description{
A function to calculate PVE by DGE, IGE, and their covariance based on the output of \code{EM3.cov()}
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
