% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\title{Calculating x-axis for Manhattan plot}
\usage{
coord(chr, pos)
}
\arguments{
\item{chr}{Chromosomal positions. Non-numeric allowed.}

\item{pos}{Genomic positions in base pairs within chromosomes.}
}
\value{
A list object including relative genomic positions and mid-points of each chromosome.
\itemize{
\item coord relative x-axis positions of Manhattan plot coefficient for self effects.
\item tic mid-point of each chromosome.
}
}
\description{
A utility function to calculate relative genomic positions to depict the x-axis of Manhattan plot
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
