% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMan.R
\name{ggMan}
\alias{ggMan}
\title{Creating Manhattan plot from GWAS output}
\usage{
ggMan(chr, pos, p)
}
\arguments{
\item{chr}{Chromosomal positions. Non-numeric allowed.}

\item{pos}{Genomic positions in base pairs within chromosomes.}

\item{p}{GWAS p-values ranging from 0 to 1.}
}
\value{
Manhattan plot generated by \code{ggplot2} package
}
\description{
A function to dipict GWAS Manhattan plot
}
\details{
For genome-wide threshold, dashed and solid lines indicate FDR and Bonferroni threshold, respectively.
The latter is provided.when FDR is not detected by \code{RAINBOWR::CalcTreshold()}.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
