% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQQ.R
\name{ggQQ}
\alias{ggQQ}
\title{Creating Manhattan plot from GWAS output}
\usage{
ggQQ(p)
}
\arguments{
\item{p}{GWAS p-values ranging from 0 to 1.}
}
\value{
QQ plot generated by \code{ggplot2} package
}
\description{
A function to dipict GWAS QQ plot
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
