% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nei.coval.R
\name{nei.coval}
\alias{nei.coval}
\title{Calculating neighbor genotypic identity}
\usage{
nei.coval(
  geno,
  smap,
  scale,
  dist.func = NULL,
  grouping = rep(1, nrow(smap)),
  n_core = 1L
)
}
\arguments{
\item{geno}{An individual x marker matrix. Bialleles (i.e., A or a) must be converted into -1 or 1 digit.}

\item{smap}{A matrix showing a spatial map for individuals. The first and second column include spatial points along an x-axis and y-axis, respectively.}

\item{scale}{A numeric scalar indicating the maximum spatial (Euclidean) distance between a focal individual and neighbors to define neighbor effects.}

\item{dist.func}{An option to set a distance decay function given as a function of scale. Default is NULL, meaning no distance decay.}

\item{grouping}{A positive integer vector assigning each individual to a group. This argument can be useful when a "smap" contains different experimental replicates. Default setting means that all individuals are belong to a single group.}

\item{n_core}{No. of cores for a multi-core computation. This does not work for Windows OS. Default is a single-core computation.}
}
\value{
A numeric matrix for neighbor covariates, with no. of individuals x markers.
}
\description{
A function to calculate neighbor genotypic identity, with a given reference scale and a degree of distance decay.
}
\details{
The argument \code{dist.func} allows a user-defined function, but default setting is recommended unless spatial distance decay of neighbor effects needs to be modeled.
If \code{dist.func} is not NULL, this argument must be pre-defined and given as a function of x. Then output variables are weighted on the basis of Euclidean distance from a focal individual up to \code{scale}.
}
\examples{
set.seed(1)
g <- matrix(sample(c(-1,1),100*1000,replace = TRUE),100,1000)
gmap <- cbind(c(rep(1,nrow(g)/2),rep(2,nrow(g)/2)),c(1:ncol(g)))
x <- rep(c(1:10),10)
y <- rep(c(1:10),each=10)
smap <- cbind(x,y)
grouping <- c(rep(1,nrow(g)/2), rep(2,nrow(g)/2))

g_nei <- nei.coval(geno=g,smap=smap,scale=sqrt(2),grouping=grouping)

# Distance decay with f(d) = 1/d^2 as suggested by Muir (2005) Genetics 170:1247–1259.
dist.func = function(x) return((1/x^2))
g_nei <- nei.coval(geno=g,smap=smap,scale=sqrt(2),dist.func=dist.func,grouping=grouping)
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
