% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neiEM3.cov.R
\name{neiEM3.cov}
\alias{neiEM3.cov}
\title{Mixed models for testing self and neighbor effects with their covariance}
\usage{
neiEM3.cov(geno, g_nei, pheno, X0 = matrix(1, nrow = length(pheno)), ...)
}
\arguments{
\item{geno}{An individual x marker matrix. Bialleles (i.e., A or a) must be converted into -1 or 1 digit.}

\item{g_nei}{An output of \code{nei_coval()} object, namely an individual x marker matrix including neighbor genotypic identity.}

\item{pheno}{A numeric vector including phenotypes for individuals}

\item{X0}{An optional matrix including additional non-genetic covariates. It contains no. of individuals x no. of covariates.}

\item{...}{Arguments to be passed to \code{EM3.cov()}.
#' @return A list object generated by \code{EM3.cov()} as well as its LL-ratio and p-value over the model without covariance (rho).}
}
\description{
A function to provide coefficients and p-values of self and neighbor effects for each marker.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp}) and Kosuke Hamazaki
}
