% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neiGWAS.cov.R
\name{neiGWAS.cov}
\alias{neiGWAS.cov}
\title{Mixed models for testing self and neighbor effects with their covariance}
\usage{
neiGWAS.cov(
  geno,
  g_nei,
  pheno,
  X0 = matrix(1, nrow = length(pheno)),
  asym = FALSE,
  n_core = 1L,
  ...
)
}
\arguments{
\item{geno}{An individual x marker matrix. Bialleles (i.e., A or a) must be converted into -1 or 1 digit.}

\item{g_nei}{An output of \code{nei_coval()} object, namely an individual x marker matrix including neighbor genotypic identity.}

\item{pheno}{A numeric vector including phenotypes for individuals}

\item{X0}{An optional matrix including additional non-genetic covariates. It contains no. of individuals x no. of covariates.}

\item{asym}{If TRUE, beta12 is also tested. Default is FALSE, which test beta2 without beta12 following the original Neighbor GWAS.}

\item{n_core}{No. of cores for a multi-core computation. This does not work for Windows OS. Default is a single-core computation.}

\item{...}{Arguments to be passed to \code{EM3.cov()}.}
}
\value{
A data.frame including coefficients and p-values of self and neighbor effects, without the chromosome numbers and marker position.
\itemize{
\item beta1 coefficient for self effects.
\item beta12 coefficient for asymmetric (one-sided) neighbor effects. Available only when asym = TRUE.
\item beta2 coefficient for symmetric (interactive) neighbor effects.
\item chisq1 likelihood ratio between a null and standard GWAS model.
\item chisq12 likelihood ratio between models with or without asymmetric neighbor effects. Available only when asym = TRUE.
\item chisq2 likelihood ratio between models with or without symmetric neighbor effects.
\item p1 p-value for self effects by a likelihood ratio test between a null and standard GWAS model.
\item p12 p-value for neighbor effects by a likelihood ratio test between models with or without asymmetric neighbor effects. Available only when asym = TRUE.
\item p2 p-value for neighbor effects by a likelihood ratio test models with or without symmetric neighbor effects.
}
}
\description{
A function to provide coefficients and p-values of self and neighbor effects for each marker.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp}) and Kosuke Hamazaki
}
