% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNeighborLMM-package.R
\docType{package}
\name{rNeighborLMM-package}
\alias{rNeighborLMM}
\alias{rNeighborLMM-package}
\title{rNeighborLMM: Fitting mixed models for oligogenic and polygenic indirect genetic effects}
\description{
This package serves as a wrapper of RAINBOWR and rNeighborGWAS, which solves a multi-kernel mixed model incorporating direct and indirect genetic effects with their covariance. GWAS of oligogenic indirect genetic effects can be performed following the same way as rNeighborGWAS operates. See Sato & Hamazaki (2026) for theoretical details.
}
\author{
\strong{Maintainer}: Yasuhiro Sato \email{sato.yasuhiro.36c@kyoto-u.jp} (\href{https://orcid.org/0000-0002-6466-723X}{ORCID})

Authors:
\itemize{
  \item Kosuke Hamazaki
}

}
\keyword{internal}
